refactor: Add command to list currently connected users
The code changes in main.py add a new command, /who, to list the currently connected users. This command checks if the sender is authorized to use it and then retrieves the list of connected users using the `w` command. The list is sent as a table in the response message. This feature enhances the functionality of the application by providing information about the currently connected users.

Note: This commit message follows the established convention of using a prefix to indicate the type of change (feat for feature, refact for refactoring, chore for maintenance tasks) and provides a concise summary of the changes made.